+++ tests/unit/test_container.py
@@ -393,7 +393,19 @@ Equation e;
         transport.solve()
         self.assertEqual(
             list(m.data.keys()),
-            ["i", "j", "a", "b", "d", "c", "x", "supply", "demand"],
+            [
+                "i",
+                "j",
+                "a",
+                "b",
+                "d",
+                "c",
+                "x",
+                "supply",
+                "demand",
+                "transport_objective_variable",
+                "transport_objective",
+            ],
         )
 
     def test_write(self):
@@ -570,98 +582,8 @@ Equation e;
             sense=Sense.MIN,
             objective=Sum((i, j), c[i, j] * x[i, j]),
         )
-        transport.solve()
 
-        # debugging level is different than keep
-        with self.assertRaises(ValidationError):
-            self.m.toGams("bla")
-
-        m = Container(debugging_level="keep")
-        i = Set(
-            m,
-            name="i",
-            records=["seattle", "san-diego"],
-            description="canning plants",
-        )
-        j = Set(
-            m,
-            name="j",
-            records=["new-york", "chicago", "topeka"],
-            description="markets",
-        )
-
-        # Data
-        a = Parameter(
-            m,
-            name="a",
-            domain=i,
-            records=self.capacities,
-            description="capacity of plant i in cases",
-        )
-        b = Parameter(
-            m,
-            name="b",
-            domain=j,
-            records=self.demands,
-            description="demand at market j in cases",
-        )
-        d = Parameter(
-            m,
-            name="d",
-            domain=[i, j],
-            records=self.distances,
-            description="distance in thousands of miles",
-        )
-        c = Parameter(
-            m,
-            name="c",
-            domain=[i, j],
-            description="transport cost in thousands of dollars per case",
-        )
-        c[i, j] = 90 * d[i, j] / 1000
-
-        # Variable
-        x = Variable(
-            m,
-            name="x",
-            domain=[i, j],
-            type="Positive",
-            description="shipment quantities in cases",
-        )
-
-        # Equation
-        supply = Equation(
-            m,
-            name="supply",
-            domain=i,
-            description="observe supply limit at plant i",
-        )
-        demand = Equation(
-            m,
-            name="demand",
-            domain=j,
-            description="satisfy demand at market j",
-        )
-
-        supply[i] = Sum(j, x[i, j]) <= a[i]
-        demand[j] = Sum(i, x[i, j]) >= b[j]
-
-        transport = Model(
-            m,
-            name="transport",
-            equations=m.getEquations(),
-            problem="LP",
-            sense=Sense.MIN,
-            objective=Sum((i, j), c[i, j] * x[i, j]),
-        )
-        transport.solve()
-
-        # File already exists
-        with self.assertRaises(ValidationError):
-            m.toGams("src")
-
-        gams_model_dir = os.path.join("tmp", "gams_model")
-        m.toGams(gams_model_dir)
+        transport.toGams(os.path.join("tmp", "to_gams"))
 
         try:
             import gamspy_base
@@ -669,7 +591,7 @@ Equation e;
             process = subprocess.run(
                 [
                     os.path.join(gamspy_base.directory, "gams"),
-                    os.path.join(gams_model_dir, "model.gms"),
+                    os.path.join("tmp", "to_gams", "transport.gms"),
                 ],
                 capture_output=True,
                 text=True,